+++ CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+### Fixed
+
+### Changed
+
+### Removed
+
+## v0.3.0
+
+### Added
+
 - Added v42 OpenAPI spec and generated API.
 - Added v42 client library support.
 - Added v41, v42 support for sdiag statistics.